image-summarisation/pyproject.toml
[project]
name = "image-summarisation"
version = "0.1.0"
description = "Summarise images at scale using vision-language models via the Doubleword Batch API"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "tqdm>=4.66.0",
    "pandas>=2.0.0",
    "pillow>=10.0.0",
    "aiohttp>=3.9.0",
    "certifi>=2023.0.0",
]

[project.scripts]
image-summarisation = "src.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
